LP-391 Changed PIOS_COM_Init buffer len parameters to uint16_t type. Fixed simposix pios_board.c GPSPORT_GPS init incorrectly passing -1 to PIOS_Board_configure_com() which was on some platforms converted to large unsigned value.

+++ flight/pios/posix/pios_com.c
@@ -118,7 +118,7 @@ static void PIOS_COM_UnblockTx(struct pios_com_dev *com_dev, bool *need_yield);
  * \param[in] id
  * \return < 0 if initialisation failed
  */
-int32_t PIOS_COM_Init(uint32_t *com_id, const struct pios_com_driver *driver, uint32_t lower_id, uint8_t *rx_buffer, size_t rx_buffer_len, uint8_t *tx_buffer, size_t tx_buffer_len)
+int32_t PIOS_COM_Init(uint32_t *com_id, const struct pios_com_driver *driver, uint32_t lower_id, uint8_t *rx_buffer, uint16_t rx_buffer_len, uint8_t *tx_buffer, uint16_t tx_buffer_len)
 {
     PIOS_Assert(com_id);
     PIOS_Assert(driver);
